Wrote HLSL generic triangle shaders

+++ res/generic_triangle/hlsl/int_triangle.hlsl.vert
@@ -1,6 +1,6 @@
 struct in_vertex {
-    [[vk::location(0)]] float3 position : POSITION;
-    [[vk::location(1)]] float4 color    : COLOR;
+    [[vk::location(0)]] int3 position : POSITION;
+    [[vk::location(1)]] float4 color  : COLOR;
 };
 
 struct out_pixel {
